requirements.txt
numpy==1.24.3
pandas==2.0.1
scikit-learn==1.2.2
matplotlib==3.7.1
seaborn==0.12.2
jupyter==1.0.0
requests==2.31.0
python-dotenv==1.0.0
torch
transformers
umap-learn
numba